The PetscSolver now uses a diffusion handler. The previous version was not actually working. SB 20141003
git-svn-id: file:///Users/MacAir/Dropbox/xolotl-cleaned/branches/diffusion_handler_branch@701 72105f29-0fe1-43bd-9363-b1e281e962a2

+++ xolotl/xolotlSolver/Monitor.cpp
@@ -244,7 +244,7 @@ PetscErrorCode computeHeliumFluence(TS ts, PetscInt timestep, PetscReal time,
 	double hx = (double) xGridLength / (PetscReal) (Mx - 1);
 
 	// The length of the time step
-	float dt = time - previousTime;
+	double dt = time - previousTime;
 
 	// Increment the fluence with the value at this current timestep
 	fluxHandler->incrementHeFluence(dt);
